refactor: move extractColor util
